+++ src/test/groovy/com/cloudogu/gitops/tools/core/JenkinsTest.groovy
@@ -89,6 +89,7 @@ me:x:1000:''')
 		assertThat(parseActualYaml()['dockerClientVersion'].toString()).isEqualTo('23')
 
 		assertThat(parseActualYaml()['controller']['image']['tag']).isEqualTo('4.8.1')
+		assertThat(parseActualYaml()['controller']['installPlugins']).isEqualTo(false)
 
 		assertThat(parseActualYaml()['controller']['jenkinsUrl']).isEqualTo('http://jenkins')
 		assertThat(parseActualYaml()['controller']['serviceType']).isEqualTo('NodePort')
@@ -119,6 +120,22 @@ me:x:1000:''')
 
 		assertThat(parseActualYaml()['agent']['runAsUser']).isEqualTo('0')
 		assertThat(parseActualYaml()['agent']['runAsGroup']).isEqualTo('133')
+	}
+
+	@Test
+	void 'Installs OIDC plugin before Jenkins startup when OIDC is configured'() {
+		config.jenkins.oidc = '''
+jenkins:
+  securityRealm:
+    oic:
+      clientId: "jenkins"
+'''
+
+		createJenkins().install()
+
+		List installedPlugins = parseActualYaml()['controller']['installPlugins'] as List
+		assertThat(installedPlugins.collect { it.toString().split(':')[0] }).containsExactly('oic-auth',
+			'json-path-api')
 	}
 
 	@Test
@@ -291,9 +308,9 @@ me:x:1000:''')
 	}
 
 	@Test
-	void 'Does not create create job credentials when argo cd is deactivated'() {
+	void 'Does not create metrics user if security realm does not support local user creation'() {
 		config.application.namePrefixForEnvVars = 'MY_PREFIX_'
-		when(userManager.isUsingCasSecurityRealm()).thenReturn(true)
+		when(userManager.isUsingSecurityRealmWithoutLocalUserCreation()).thenReturn(true)
 
 		createJenkins().install()
